requirements.txt
flask
psycopg2-binary
gunicorn
python-dotenv
google-genai